move code to utils/file
