+++ app/components/ui-lib.tsx
@@ -511,29 +511,3 @@ export function Selector<T>(props: {
     </div>
   );
 }
-
-export function IndexDBImage({ src, alt, onClick, db, className }) {
-  const [data, setData] = useState(src);
-  const imgId = useMemo(
-    () => src.replace("indexeddb://", "").split("@").pop(),
-    [src],
-  );
-  useEffect(() => {
-    db.getByID(imgId)
-      .then(({ data }) => {
-        setData(`data:image/png;base64,${data}`);
-      })
-      .catch((e) => {
-        setData(src);
-      });
-  }, [src, imgId]);
-
-  return (
-    <img
-      className={className}
-      src={data}
-      alt={alt}
-      onClick={(e) => onClick(data, e)}
-    />
-  );
-}
